Renamed selection-to-view-converters.js to model-selection-to-view-converters.js.
