+++ src/treecontroller/model-converter-builder.js
@@ -13,7 +13,7 @@ import {
 	unwrap
 } from '/ckeditor5/engine/treecontroller/model-to-view-converters.js';
 
-import { convertSelectionAttribute } from '/ckeditor5/engine/treecontroller/selection-to-view-converters.js';
+import { convertSelectionAttribute } from '/ckeditor5/engine/treecontroller/model-selection-to-view-converters.js';
 
 import ViewAttributeElement from '/ckeditor5/engine/treeview/attributeelement.js';
 import ViewContainerElement from '/ckeditor5/engine/treeview/containerelement.js';
