+++ src/treecontroller/model-selection-to-view-converters.js
